backend/services/notification_alerting/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
celery
redis
smtplib
email
